+++ armsrc/appmain.h
@@ -44,8 +44,8 @@ void send_wtx(uint16_t wtx);
 void ReadMem(int addr);
 void __attribute__((noreturn)) AppMain(void);
 
-uint16_t AvgAdc(int ch);
-uint16_t SumAdc(int ch, int NbSamples);
+uint16_t AvgAdc(uint8_t ch);
+uint16_t SumAdc(uint8_t ch, uint8_t NbSamples);
 
 //void PrintToSendBuffer(void);
 void ToSendStuffBit(int b);
